Update  - SPO engine  - Refactoring code
