+++ Cargo.toml
@@ -23,4 +23,5 @@ futures = "0.3"
 uuid = { version =  "1.6", default-features = false , features = ["v4","fast-rng","macro-diagnostics"] }
 console = "0.15"
 spinner = "0.5"
+time = "0.3"
